requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
asyncpg==0.30.0
exceptiongroup==1.2.2
fastapi==0.115.3
greenlet==3.1.1
idna==3.10
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.0
typing_extensions==4.12.2